Sample MP3 files
The filenames of the mp3 files needs to be in the correct order as it is use in the Arduino sketch or code. I recommend to use FAT sorter when writing the files to the SD card for DF player module. You have to compress the mp3 files to lower bit rate to be able to use the internal memory of the JQ6500 mp3 module.